refactor: make test w client hinerit same class
(so they can all use same fixture)
